renaming micro, adding register endpoint, will implment hashing next
