--- backend/account/.gitignore
@@ -1,5 +1,0 @@
-.vscode
-accountmicro
-.env
-build
-coverage.out
