--- backend/account/internal/dtos/account_dto.go
@@ -1,6 +1,0 @@
-package dto
-
-type AccountDTO struct {
-	Email    *string `json:"email"`
-	Password *string `json:"password"`
-}
