--- backend/account/internal/error/customerrors/global_database_error.go
@@ -1,18 +1,0 @@
-package customerrors
-
-type GlobalDatabaseError struct {
-	Message string
-	Code    int
-}
-
-func (e *GlobalDatabaseError) Error() string {
-	return e.Message
-}
-
-func (e *GlobalDatabaseError) Is(tgt error) bool {
-	target, ok := tgt.(*GlobalDatabaseError)
-	if !ok {
-		return false
-	}
-	return e.Code == target.Code
-}
