--- backend/account/internal/error/customerrors/resource_conflict_error.go
@@ -1,18 +1,0 @@
-package customerrors
-
-type ResourceConflictError struct {
-	Message string
-	Code    int
-}
-
-func (e *ResourceConflictError) Error() string {
-	return e.Message
-}
-
-func (e *ResourceConflictError) Is(tgt error) bool {
-	target, ok := tgt.(*ResourceConflictError)
-	if !ok {
-		return false
-	}
-	return e.Code == target.Code
-}
