--- backend/account/internal/models/account_model.go
@@ -1,10 +1,0 @@
-package models
-
-import "github.com/google/uuid"
-
-type Account struct {
-	ID           uuid.UUID `json:"id" gorm:"type:uuid;primaryKey"`
-	Email        *string   `json:"email"`
-	Password     *string   `json:"password"`
-	RefreshToken *string   `json:"refresh"`
-}
