+++ backend/auth/.gitignore
@@ -1,0 +1,5 @@
+.vscode
+accountmicro
+.env
+build
+coverage.out
