+++ backend/auth/internal/dtos/auth_dto.go
@@ -1,0 +1,6 @@
+package dto
+
+type AuthDTO struct {
+	Email    *string `json:"email"`
+	Password *string `json:"password"`
+}
