+++ backend/auth/internal/error/customerrors/bad_request_error.go
@@ -1,0 +1,18 @@
+package customerrors
+
+type BadRequestError struct {
+	Message string
+	Code    int
+}
+
+func (e *BadRequestError) Error() string {
+	return e.Message
+}
+
+func (e *BadRequestError) Is(tgt error) bool {
+	target, ok := tgt.(*BadRequestError)
+	if !ok {
+		return false
+	}
+	return e.Code == target.Code
+}
