+++ go.work
@@ -1,6 +1,6 @@
 go 1.21.1
 
 use (
-	./backend/account
+	./backend/auth
 	./backend/prescription
 )
